Hasan_AlTalib_A23MJ0018/ml-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
scikit-learn>=1.5.2,<2.0
lightgbm==4.6.0
fastembed==0.4.1
pymongo==4.10.1
python-dotenv==1.0.1
joblib==1.4.2
numpy==1.26.4
scipy==1.14.1
rank-bm25==0.2.2
